refactor: split board and prewarm logic
