+++ src/main.ts
@@ -1,28 +1,24 @@
 // src/main.ts
-import { Plugin, Notice, Modal as ObsidianModal, Hotkey, Modifier, Setting, App } from 'obsidian';
+import { Plugin } from 'obsidian';
 import type { PluginGlobalSettings, BoardConfiguration, WidgetConfig } from './interfaces';
 import { DEFAULT_PLUGIN_SETTINGS, DEFAULT_BOARD_CONFIGURATION } from './settingsDefaults';
-import { WidgetBoardModal } from './modal';
 import { WidgetBoardSettingTab } from './settingsTab';
 import { registeredWidgetImplementations } from './widgetRegistry';
 import { DEFAULT_POMODORO_SETTINGS, PomodoroSettings } from './widgets/pomodoro';
-import { DEFAULT_MEMO_SETTINGS, MemoWidgetSettings } from './widgets/memo';
+import { DEFAULT_MEMO_SETTINGS } from './widgets/memo';
 import { DEFAULT_CALENDAR_SETTINGS } from './settingsDefaults';
 import { DEFAULT_TIMER_STOPWATCH_SETTINGS, TimerStopwatchWidgetSettings } from './widgets/timer-stopwatch';
 import cloneDeep from 'lodash.clonedeep';
 import { LLMManager } from './llm/llmManager';
 import { GeminiProvider } from './llm/gemini/geminiApi';
 import yaml from 'js-yaml';
-import { TweetRepository } from './widgets/tweetWidget';
-import { renderMarkdownBatchWithCache } from './utils/renderMarkdownBatch';
 import { debugLog } from './utils/logger';
 import { filterConsoleWarn } from './utils/consoleWarnFilter';
-import { Component, TFile } from 'obsidian';
 import { preloadChartJS } from './widgets/reflectionWidget/reflectionWidgetUI';
-import { getDateKey, getWeekRange } from './utils';
-import { FileViewWidgetSettings } from './widgets/file-view';
 import { TweetWidgetSettings } from './widgets/tweetWidget/types';
 import { ReflectionWidgetSettings } from './widgets/reflectionWidget/reflectionWidgetTypes';
+import { BoardManager } from './boardManager';
+import { PrewarmManager } from './prewarm';
 
 /**
  * Obsidian Widget Board Pluginのメインクラス
@@ -31,12 +27,12 @@ import { ReflectionWidgetSettings } from './widgets/reflectionWidget/reflectionW
 export default class WidgetBoardPlugin extends Plugin {
     /** プラグイン全体の設定 */
     settings: PluginGlobalSettings;
-    /** 開いているウィジェットボードのモーダル管理 */
-    widgetBoardModals: Map<string, WidgetBoardModal> = new Map();
+    /** ボード管理 */
+    boardManager: BoardManager;
+    /** プリウォーム・統計管理 */
+    prewarmManager: PrewarmManager;
     private isSaving: boolean = false;
-    private registeredGroupCommandIds: string[] = [];
     llmManager: LLMManager;
-    tweetPostCountCache: Record<string, number> = {};
     tweetChartDirty: boolean = true;
     tweetChartImageData: string | null = null;
     tweetChartCountsKey: string | null = null;
@@ -50,11 +46,13 @@ export default class WidgetBoardPlugin extends Plugin {
         filterConsoleWarn(['[Violation]', '[Deprecation]']);
         this.llmManager = new LLMManager();
         this.llmManager.register(GeminiProvider);
+        this.boardManager = new BoardManager(this);
+        this.prewarmManager = new PrewarmManager(this);
         // Preload Chart.js for ReflectionWidget without blocking startup
         setTimeout(() => preloadChartJS().catch(() => {}), 0);
         await this.loadSettings();
-        await this.initTweetPostCountCache();
-        this.registerAllBoardCommands();
+        await this.prewarmManager.initTweetPostCountCache();
+        this.boardManager.registerAllBoardCommands();
         this.addRibbonIcon('layout-dashboard', 'ウィジェットボードを開く', () => this.openBoardPicker());
         this.addSettingTab(new WidgetBoardSettingTab(this.app, this));
         // すべてのボードをトグルで非表示/表示するコマンドを追加
@@ -127,7 +125,7 @@ export default class WidgetBoardPlugin extends Plugin {
             }
         );
         // プリウォーム処理を追加
-        this.prewarmAllWidgetMarkdownCache();
+        this.prewarmManager.prewarmAllWidgetMarkdownCache();
         debugLog(this, 'Widget Board Plugin: Loaded.');
     }
 
@@ -136,93 +134,27 @@ export default class WidgetBoardPlugin extends Plugin {
      * @override
      */
     onunload(): void {
-        this.widgetBoardModals.forEach(modal => {
-            if (modal.isOpen) modal.close();
-        });
-        this.widgetBoardModals.clear();
+        this.boardManager.hideAllBoards();
+        this.boardManager.widgetBoardModals.clear();
         debugLog(this, 'Widget Board Plugin: Unloaded.');
     }
 
-    /**
-     * ボード選択モーダルを表示
-     */
     openBoardPicker(): void {
-        if (this.settings.boards.length === 0) {
-            new Notice('設定されているウィジェットボードがありません。設定画面から作成してください。');
-            return;
-        }
-        if (this.settings.boards.length === 1) {
-            this.openWidgetBoardById(this.settings.boards[0].id);
-            return;
-        }
-        const modal = new BoardPickerModal(this.app, this.settings.boards, (boardId) => {
-            this.openWidgetBoardById(boardId);
-        });
-        modal.open();
+        this.boardManager.openBoardPicker();
     }
 
-    /**
-     * 指定IDのウィジェットボードを開く
-     * @param boardId ボードID
-     */
     openWidgetBoardById(boardId: string): void {
-        const modal = this.widgetBoardModals.get(boardId);
-        if (modal) {
-            if (modal.isOpen || modal.isClosing) {
-                // 既に開いている、または閉じるアニメーション中なら何もしない
-                return;
-            } else {
-                // isOpen=false だがMapに残っている場合は削除
-                this.widgetBoardModals.delete(boardId);
-                if (modal.modalEl && modal.modalEl.parentElement === document.body) {
-                    modal.modalEl.parentElement.removeChild(modal.modalEl);
-                }
-            }
-        }
-        const boardConfig = this.settings.boards.find(b => b.id === boardId);
-        if (!boardConfig) {
-            new Notice(`ID '${boardId}' のウィジェットボードが見つかりません。`);
-            return;
-        }
-        const validModes: string[] = Object.values(WidgetBoardModal.MODES);
-        if (!validModes.includes(boardConfig.defaultMode)) {
-            boardConfig.defaultMode = WidgetBoardModal.MODES.RIGHT_THIRD;
-        }
-        // lastOpenedBoardIdやsaveSettingsは明示的な設定変更時のみ更新
-        const newModal = new WidgetBoardModal(this.app, this, boardConfig);
-        this.widgetBoardModals.set(boardId, newModal);
-        // モーダルが閉じられたらMapから削除
-        const origOnClose = newModal.onClose.bind(newModal);
-        newModal.onClose = () => {
-            this.widgetBoardModals.delete(boardId);
-            origOnClose();
-        };
-        newModal.open();
+        this.boardManager.openWidgetBoardById(boardId);
     }
 
-    /**
-     * 指定IDのウィジェットボードをトグル表示
-     * @param boardId ボードID
-     */
     toggleWidgetBoardById(boardId: string): void {
-        const modal = this.widgetBoardModals.get(boardId);
-        if (modal && modal.isOpen) {
-            modal.close(); // このモーダルだけ閉じる
-            return;
-        }
-        this.openWidgetBoardById(boardId); // このモーダルだけ開く
+        this.boardManager.toggleWidgetBoardById(boardId);
     }
 
-    /**
-     * 指定IDのウィジェットボードを閉じる
-     * @param boardId ボードID
-     */
     closeWidgetBoardById(boardId: string): void {
-        const modal = this.widgetBoardModals.get(boardId);
-        if (modal && modal.isOpen) {
-            modal.close();
-        }
+        this.boardManager.closeWidgetBoardById(boardId);
     }
+
 
     /**
      * 設定をロード（旧形式からのマイグレーションも対応）
@@ -284,7 +216,7 @@ export default class WidgetBoardPlugin extends Plugin {
         try {
             await this.saveData(this.settings);
             if (targetBoardId) {
-                const modal = this.widgetBoardModals.get(targetBoardId);
+                const modal = this.boardManager.widgetBoardModals.get(targetBoardId);
                 if (modal && modal.isOpen && modal.currentBoardConfig) {
                     const updatedBoardConfig = this.settings.boards.find(b => b.id === targetBoardId);
                     if (updatedBoardConfig) {
@@ -295,7 +227,7 @@ export default class WidgetBoardPlugin extends Plugin {
                 }
             } else {
                 // 引数なしの場合は従来通り全モーダル更新
-                this.widgetBoardModals.forEach(modal => {
+                this.boardManager.widgetBoardModals.forEach(modal => {
                     if (modal.isOpen && modal.currentBoardConfig) {
                         const currentBoardId = modal.currentBoardConfig.id;
                         const updatedBoardConfig = this.settings.boards.find(b => b.id === currentBoardId);
@@ -316,241 +248,26 @@ export default class WidgetBoardPlugin extends Plugin {
      * すべてのボード・グループコマンドを再登録
      */
     registerAllBoardCommands(): void {
-        // 既存のグループコマンドを一旦解除
-        if (typeof this.removeCommand === 'function') {
-            this.registeredGroupCommandIds.forEach(id => {
-                this.removeCommand(id);
-            });
-        }
-        this.registeredGroupCommandIds = [];
-        // ボードごとにコマンドを動的登録
-        this.settings.boards.map(board => {
-            this.addCommand({
-                id: `toggle-widget-board-${board.id}`,
-                name: `ウィジェットボードをトグル: ${board.name}`,
-                callback: () => this.toggleWidgetBoardById(board.id)
-            });
-            this.addCommand({
-                id: `close-widget-board-${board.id}`,
-                name: `ウィジェットボードを閉じる: ${board.name}`,
-                callback: () => this.closeWidgetBoardById(board.id)
-            });
-        });
-        // ボードグループごとにコマンドを動的登録
-        (this.settings.boardGroups || []).forEach(group => {
-            let hotkeys: Hotkey[] = [];
-            if (group.hotkey) {
-                const parts = group.hotkey.split('+');
-                if (parts.length > 1) {
-                    hotkeys = [{ modifiers: parts.slice(0, -1) as Modifier[], key: parts[parts.length - 1] }];
-                } else {
-                    hotkeys = [{ modifiers: [], key: group.hotkey }];
-                }
-            }
-            const cmdId = `open-board-group-${group.id}`;
-            this.addCommand({
-                id: cmdId,
-                name: `ボードグループをトグル: ${group.name}`,
-                hotkeys,
-                callback: () => {
-                    const anyOpen = (group.boardIds || []).some(boardId => {
-                        const modal = this.widgetBoardModals.get(boardId);
-                        return modal && modal.isOpen;
-                    });
-                    if (anyOpen) {
-                        (group.boardIds || []).filter(boardId => {
-                            const modal = this.widgetBoardModals.get(boardId);
-                            return modal && modal.isOpen;
-                        }).forEach(boardId => {
-                            const modal = this.widgetBoardModals.get(boardId);
-                            if (modal) modal.close();
-                        });
-                    } else {
-                        (group.boardIds || []).forEach(boardId => {
-                            this.openWidgetBoardById(boardId);
-                        });
-                    }
-                }
-            });
-            this.registeredGroupCommandIds.push(cmdId);
-        });
+        this.boardManager.registerAllBoardCommands();
     }
 
-    /**
-     * すべてのウィジェットボードを非表示にする
-     */
     hideAllBoards(): void {
-        this.widgetBoardModals.forEach(modal => {
-            if (modal.isOpen) modal.close();
-        });
+        this.boardManager.hideAllBoards();
     }
 
     private async initTweetPostCountCache() {
-        const dbPath = this.settings.baseFolder
-            ? `${this.settings.baseFolder.replace(/\/$/, '')}/tweets.json`
-            : 'tweets.json';
-        const repo = new TweetRepository(this.app, dbPath);
-        const tweetSettings = await repo.load();
-        this.tweetPostCountCache = {};
-        for (const p of tweetSettings.posts || []) {
-            if (p.deleted) continue;
-            const key = getDateKey(new Date(p.created));
-            this.tweetPostCountCache[key] = (this.tweetPostCountCache[key] || 0) + 1;
-        }
+        await this.prewarmManager.initTweetPostCountCache();
     }
 
     public updateTweetPostCount(created: number, delta: number) {
-        const key = getDateKey(new Date(created));
-        this.tweetPostCountCache[key] = (this.tweetPostCountCache[key] || 0) + delta;
-        if (this.tweetPostCountCache[key] <= 0) {
-            delete this.tweetPostCountCache[key];
-        }
-        this.tweetChartDirty = true;
+        this.prewarmManager.updateTweetPostCount(created, delta);
     }
 
     public getTweetPostCounts(days: string[]): number[] {
-        return days.map(d => this.tweetPostCountCache[d] || 0);
+        return this.prewarmManager.getTweetPostCounts(days);
     }
 
-    /**
-     * すべてのウィジェットのMarkdownをプリウォームしてキャッシュを作成
-     */
     private async prewarmAllWidgetMarkdownCache() {
-        const MAX_PREWARM_ENTRIES = 50;
-        try {
-            new Notice('キャッシュ中…');
-            // --- TweetWidget ---
-            const dbPath = this.settings.baseFolder
-                ? `${this.settings.baseFolder.replace(/\/$/, '')}/tweets.json`
-                : 'tweets.json';
-            const repo = new TweetRepository(this.app, dbPath);
-            const tweetSettings = await repo.load();
-            const tweetPosts = (tweetSettings.posts || []).slice(0, MAX_PREWARM_ENTRIES);
-
-            // --- MemoWidget, FileViewWidget ---
-            const memoContents: string[] = [];
-            const fileViewFiles: string[] = [];
-            for (const board of this.settings.boards) {
-                for (const widget of board.widgets) {
-                    if (widget.type === 'memo') {
-                        const settings = widget.settings as MemoWidgetSettings;
-                        if (settings?.memoContent) {
-                            memoContents.push(settings.memoContent);
-                        }
-                    }
-                    if (widget.type === 'file-view-widget') {
-                        const settings = widget.settings as FileViewWidgetSettings;
-                        if (settings?.fileName) {
-                            fileViewFiles.push(settings.fileName);
-                        }
-                    }
-                }
-            }
-            memoContents.splice(MAX_PREWARM_ENTRIES);
-            fileViewFiles.splice(MAX_PREWARM_ENTRIES);
-
-            // --- ReflectionWidget: AI要約（今日・今週） ---
-            async function loadReflectionSummary(type: 'today' | 'week', dateKey: string, app: App): Promise<string | null> {
-                const path = 'data.json';
-                try {
-                    const raw = await app.vault.adapter.read(path);
-                    const data = JSON.parse(raw);
-                    if (data.reflectionSummaries && data.reflectionSummaries[type]?.date === dateKey) {
-                        return data.reflectionSummaries[type].summary;
-                    }
-                } catch {
-                    // ignore JSON parse errors or missing files
-                }
-                return null;
-            }
-            const todayKey = getDateKey(new Date());
-            const [, weekEnd] = getWeekRange(this.settings.weekStartDay);
-            const weekKey = weekEnd;
-            const todaySummary = await loadReflectionSummary('today', todayKey, this.app);
-            const weekSummary = await loadReflectionSummary('week', weekKey, this.app);
-
-            // プリウォーム対象をまとめてバッチ処理
-            let tweetIndex = 0, memoIndex = 0, fileIndex = 0;
-            let reflectionIndex = 0;
-            const reflectionSummaries = [todaySummary, weekSummary].filter(Boolean).slice(0, MAX_PREWARM_ENTRIES) as string[];
-            const batchSize = 3;
-                const schedule = (cb: () => void) => {
-                    const w = window as Window & {
-                        requestIdleCallback?: (callback: IdleRequestCallback) => number;
-                    };
-                    if (typeof w.requestIdleCallback === 'function') {
-                        w.requestIdleCallback(cb);
-                    } else {
-                        requestAnimationFrame(cb);
-                    }
-                };
-            const processBatch = async () => {
-                // TweetWidget
-                const tweetEnd = Math.min(tweetIndex + batchSize, tweetPosts.length);
-                for (; tweetIndex < tweetEnd; tweetIndex++) {
-                    const post = tweetPosts[tweetIndex];
-                    await renderMarkdownBatchWithCache(post.text, document.createElement('div'), '', new Component());
-                }
-                // MemoWidget
-                const memoEnd = Math.min(memoIndex + batchSize, memoContents.length);
-                for (; memoIndex < memoEnd; memoIndex++) {
-                    await renderMarkdownBatchWithCache(memoContents[memoIndex], document.createElement('div'), '', new Component());
-                }
-                // FileViewWidget
-                const fileEnd = Math.min(fileIndex + batchSize, fileViewFiles.length);
-                for (; fileIndex < fileEnd; fileIndex++) {
-                    const file = this.app.vault.getAbstractFileByPath(fileViewFiles[fileIndex]);
-                    if (file && file instanceof TFile) {
-                        const content = await this.app.vault.read(file);
-                        await renderMarkdownBatchWithCache(content, document.createElement('div'), file.path, new Component());
-                    }
-                }
-                // ReflectionWidget AI要約
-                const reflectionEnd = Math.min(reflectionIndex + batchSize, reflectionSummaries.length);
-                for (; reflectionIndex < reflectionEnd; reflectionIndex++) {
-                    await renderMarkdownBatchWithCache(reflectionSummaries[reflectionIndex], document.createElement('div'), '', new Component());
-                }
-                if (
-                    tweetIndex < tweetPosts.length ||
-                    memoIndex < memoContents.length ||
-                    fileIndex < fileViewFiles.length ||
-                    reflectionIndex < reflectionSummaries.length
-                ) {
-                    schedule(processBatch);
-                } else {
-                    new Notice('キャッシュ完了');
-                }
-            };
-            schedule(processBatch);
-        } catch (e) {
-            console.error('プリウォーム中にエラー:', e);
-        }
-    }
-}
-
-class BoardPickerModal extends ObsidianModal {
-    constructor(
-        app: App,
-        private boards: BoardConfiguration[],
-        private onChoose: (boardId: string) => void
-    ) {
-        super(app);
-    }
-    onOpen() {
-        const { contentEl } = this;
-        contentEl.empty();
-        new Setting(contentEl).setName('ウィジェットボードを選択').setHeading();
-        const listEl = contentEl.createDiv({ cls: 'widget-board-picker-list' });
-        this.boards.forEach(board => {
-            const boardItemEl = listEl.createDiv({ cls: 'widget-board-picker-item' });
-            boardItemEl.setText(board.name);
-            boardItemEl.onClickEvent(() => {
-                this.onChoose(board.id);
-                this.close();
-            });
-        });
-    }
-    onClose() {
-        this.contentEl.empty();
+        await this.prewarmManager.prewarmAllWidgetMarkdownCache();
     }
 }
